+++ tests/test_main.py
@@ -5,6 +5,37 @@ import pytest
 
 from extremitypathfinder.extremitypathfinder import PolygonEnvironment
 
+# size_x, size_y, obstacle_iter
+GRID_ENV_PARAMS = (19, 10, [
+    # (x,y),
+    # obstacles changing boundary
+    (0, 1),
+    (1, 1),
+    (2, 1),
+    (3, 1),
+
+    (17, 9),
+    (17, 8),
+    (17, 7),
+
+    (17, 5),
+    (17, 4),
+    (17, 3),
+    (17, 2),
+    (17, 1),
+    (17, 0),
+
+    # hole 1
+    (5, 5),
+    (5, 6),
+    (6, 6),
+    (6, 7),
+    (7, 7),
+
+    # hole 2
+    (7, 5),
+])
+
 INVALID_DESTINATION_DATA = [
     # outside of map region
     ((-1, 5.0), (17, 0.5)),
@@ -27,33 +58,28 @@ INVALID_DESTINATION_DATA = [
     ((17, 0.5), (6.5, 6.5)),
 ]
 
-PATH_TEST_DATA = [
+TEST_DATA_GRID_ENV = [
     # ((start,goal),(path,distance))
+    # shortest paths should be distinct (reverse will automatically be tested)
+
     # identical nodes
     (((15, 5), (15, 5)), ([(15, 5), (15, 5)], 0.0)),
 
     # directly reachable
     (((15, 5), (15, 6)), ([(15, 5), (15, 6)], 1.0)),
-    (((15, 6), (15, 5)), ([(15, 6), (15, 5)], 1.0)),
     (((15, 5), (16, 6)), ([(15, 5), (16, 6)], sqrt(2))),
-    (((16, 6), (15, 5)), ([(16, 6), (15, 5)], sqrt(2))),
 
     # points on the polygon edges (vertices) should be accepted!
     # on edge
     (((15, 0), (15, 6)), ([(15, 0), (15, 6)], 6.0)),
-    (((15, 6), (15, 0)), ([(15, 6), (15, 0)], 6.0)),
     (((17, 5), (16, 5)), ([(17, 5), (16, 5)], 1.0)),
-    (((16, 5), (17, 5)), ([(16, 5), (17, 5)], 1.0)),
     # on edge of hole
     (((7, 8), (7, 9)), ([(7, 8), (7, 9)], 1.0)),
-    (((7, 9), (7, 8)), ([(7, 9), (7, 8)], 1.0)),
 
     # on vertex
     (((4, 2), (4, 3)), ([(4, 2), (4, 3)], 1.0)),
-    (((4, 3), (4, 2)), ([(4, 3), (4, 2)], 1.0)),
     # on vertex of hole
     (((6, 8), (6, 9)), ([(6, 8), (6, 9)], 1.0)),
-    (((6, 9), (6, 8)), ([(6, 9), (6, 8)], 1.0)),
 
     # on two vertices
     # coinciding with edge (direct neighbour)
@@ -91,12 +117,11 @@ PATH_TEST_DATA = [
     (((0.5, 6), (18.5, 0.5)),
      ([(0.5, 6.0), (5, 5), (6, 5), (7, 5), (8, 5), (17, 6), (18, 6), (18.5, 0.5)], 23.18783787537749)),
 
-    (((0.5, 6), (9, 6)),
-     ([(0.5, 6.0), (5, 5), (6, 5), (7, 6), (8, 6), (9, 6)], 9.023985791019538)),
+    (((0.5, 6), (9, 5.5)),
+     ([(0.5, 6.0), (5, 5), (6, 5), (7, 5), (8, 5), (9.0, 5.5)], 8.727806217396338)),
 
     (((0.5, 6), (18.5, 9)),
      ([(0.5, 6.0), (5, 5), (6, 5), (7, 5), (8, 5), (18, 7), (18.5, 9.0)], 19.869364068640845)),
-    # 19.86936406864084
 
     (((6.9, 4), (7, 9)),
      ([(6.9, 4.0), (7, 6), (8, 7), (8, 8), (7, 9)], 5.830925564196269)),
@@ -105,90 +130,142 @@ PATH_TEST_DATA = [
      ([(6.5, 4.0), (7, 6), (8, 7), (8, 8), (7, 9)], 5.889979937555021)),
 ]
 
+POLY_ENV_PARAMS = (
+    # boundary_coordinates
+    [(0.0, 0.0), (10.0, 0.0), (9.0, 5.0), (10.0, 10.0), (0.0, 10.0)],
+    # list_of_holes
+    [[(3.0, 7.0), (5.0, 9.0), (4.6, 7.0), (5.0, 4.0), ], ]
+)
+
+TEST_DATA_POLY_ENV = [
+    # ((start,goal),(path,distance))
+    # identical nodes
+    (((1, 1), (1, 1)), ([(1, 1), (1, 1)], 0.0)),
+
+    # directly reachable
+    (((1, 1), (1, 2)), ([(1, 1), (1, 2)], 1.0)),
+    (((1, 1), (2, 1)), ([(1, 1), (2, 1)], 1.0)),
+
+    # points on the polygon edges (vertices) should be accepted!
+    # on edge (boundary polygon)
+    (((1, 0), (1, 1)), ([(1, 0), (1, 1)], 1.0)),
+    (((9.5, 2.5), (8.5, 2.5)), ([(9.5, 2.5), (8.5, 2.5)], 1.0)),
+    (((0, 2), (0, 1)), ([(0, 2), (0, 1)], 1.0)),  # both
+    (((1, 0), (5, 0)), ([(1, 0), (5, 0)], 4.0)),  # both
+    # on edge of hole
+    (((4, 8), (3, 8)), ([(4, 8), (3, 8)], 1.0)),
+    (((4, 8), (4.1, 8.1)), ([(4, 8), (4.1, 8.1)], sqrt(2 * (0.1 ** 2)))),  # both
+
+    # on vertex
+    (((9, 5), (8, 5)), ([(9, 5), (8, 5)], 1.0)),
+    # on vertex of hole
+    (((3, 7), (2, 7)), ([(3, 7), (2, 7)], 1.0)),
+
+    # on two vertices
+    # coinciding with edge (direct neighbour)
+    (((3, 7), (5, 9)), ([(3, 7), (5, 9)], sqrt(8))),
+    (((4.6, 7), (5, 9)), ([(4.6, 7), (5, 9)], sqrt((0.4 ** 2) + (2 ** 2)))),
+    # should have direct connection to all visible extremities! connected in graph
+    (((5, 4), (5, 9)), ([(5, 4), (5, 9)], 5)),
+    # should have a direct connection to all visible extremities! even if not connected in graph!
+    (((9, 5), (5, 9)), ([(9, 5), (5, 9)], sqrt(2 * (4 ** 2)))),
+
+    # using a* graph search:
+    # directly reachable through a single vertex (does not change distance!)
+    (((9, 4), (9, 6)), ([(9, 4), (9, 5), (9, 6)], 2)),
+
+    # # If two Polygons have vertices with identical coordinates (this is allowed),
+    # #   paths through these vertices are theoretically possible!
+    # (((6.5, 5.5), (7.5, 6.5)), ([(6.5, 5.5), (7, 6), (7.5, 6.5)], sqrt(1 ** 2 + 1 ** 2))),
+    #
+    # # distance should stay the same even if multiple extremities lie on direct path
+    # # test if path is skipping passed extremities
+    # (((8, 4), (8, 8)), ([(8, 4), (8, 5), (8, 6), (8, 7), (8, 8)], 4)),
+    # (((8, 4), (8, 9)), ([(8, 4), (8, 5), (8, 6), (8, 7), (8, 8), (8, 9)], 5)),
+    #
+    # # regular examples
+    # (((0.5, 6), (18.5, 0.5)),
+    #  ([(0.5, 6.0), (5, 5), (6, 5), (7, 5), (8, 5), (17, 6), (18, 6), (18.5, 0.5)], 23.18783787537749)),
+    #
+    # (((0.5, 6), (9, 6)),
+    #  ([(0.5, 6.0), (5, 5), (6, 5), (7, 6), (8, 6), (9, 6)], 9.023985791019538)),
+    #
+    # (((0.5, 6), (18.5, 9)),
+    #  ([(0.5, 6.0), (5, 5), (6, 5), (7, 5), (8, 5), (18, 7), (18.5, 9.0)], 19.869364068640845)),
+    #
+    # (((6.9, 4), (7, 9)),
+    #  ([(6.9, 4.0), (7, 6), (8, 7), (8, 8), (7, 9)], 5.830925564196269)),
+    #
+    # (((6.5, 4), (7, 9)),
+    #  ([(6.5, 4.0), (7, 6), (8, 7), (8, 8), (7, 9)], 5.889979937555021)),
+]
+
+
+def try_test_cases(environment, test_cases):
+    def validate(start_coordinates, goal_coordinates, expected_output):
+        output = environment.find_shortest_path(start_coordinates, goal_coordinates)
+        path, length = output
+        expected_path, expected_length = expected_output
+        correct_result = path == expected_path and length == pytest.approx(expected_length)
+        if correct_result:
+            status_str = 'OK'
+        else:
+            status_str = 'XX'
+        print(f'{status_str} input: {(start_coordinates, goal_coordinates)} ')
+        assert correct_result, \
+            f'unexpected result (path, length): got {output} instead of {expected_output} '
+
+    print('testing if path and distance are correct:')
+    for ((start_coordinates, goal_coordinates), expected_output) in test_cases:
+        validate(start_coordinates, goal_coordinates, expected_output)
+        # automatically test reversed!
+        path, length = expected_output
+        expected_output_reversed = list(reversed(path)), length
+        validate(goal_coordinates, start_coordinates, expected_output_reversed)
+
 
 class MainTest(unittest.TestCase):
 
     def test_fct(self):
-        environment = PolygonEnvironment()
+        grid_env = PolygonEnvironment()
 
-        size_x, size_y = 19, 10
-        obstacle_iter = [
-            # (x,y),
-
-            # obstacles changing boundary
-            (0, 1),
-            (1, 1),
-            (2, 1),
-            (3, 1),
-
-            (17, 9),
-            (17, 8),
-            (17, 7),
-
-            (17, 5),
-            (17, 4),
-            (17, 3),
-            (17, 2),
-            (17, 1),
-            (17, 0),
-
-            # hole 1
-            (5, 5),
-            (5, 6),
-            (6, 6),
-            (6, 7),
-            (7, 7),
-
-            # hole 2
-            (7, 5),
-        ]
-
-        environment.store_grid_world(size_x, size_y, obstacle_iter, simplify=False, validate=False)
-        assert len(environment.all_extremities) == 17, 'extremities do not get detected correctly!'
-        environment.prepare()
-        assert len(environment.graph.all_nodes) == 16, 'identical nodes should get joined in the graph!'
+        grid_env.store_grid_world(*GRID_ENV_PARAMS, simplify=False, validate=False)
+        assert len(grid_env.all_extremities) == 17, 'extremities do not get detected correctly!'
+        grid_env.prepare()
+        # raise ValueError
+        assert len(grid_env.graph.all_nodes) == 16, 'identical nodes should get joined in the graph!'
 
         # test if points outside the map are being rejected
         for start_coordinates, goal_coordinates in INVALID_DESTINATION_DATA:
             with pytest.raises(ValueError):
-                environment.find_shortest_path(start_coordinates, goal_coordinates)
+                grid_env.find_shortest_path(start_coordinates, goal_coordinates)
 
-        print('testing if path and distance are correct:')
-        for ((start_coordinates, goal_coordinates), expected_output) in PATH_TEST_DATA:
-            actual_output = environment.find_shortest_path(start_coordinates, goal_coordinates)
-            correct_result = actual_output == expected_output
-            if correct_result:
-                status_str = 'OK'
-            else:
-                status_str = 'XX'
-            print(f'{status_str} input: {(start_coordinates, goal_coordinates)} ')
-            assert correct_result, \
-                f'unexpected result (path, length): got {actual_output} instead of {expected_output} '
+        try_test_cases(grid_env, TEST_DATA_GRID_ENV)
 
         # when the deep copy mechanism works correctly
         # even after many queries the internal graph should have the same structure as before
         # otherwise the temporarily added vertices during a query stay stored
-        assert len(environment.graph.all_nodes) == 16, 'the graph should stay unchanged by shortest path queries!'
+        assert len(grid_env.graph.all_nodes) == 16, 'the graph should stay unchanged by shortest path queries!'
 
-        nr_nodes_env1_old = len(environment.graph.all_nodes)
+        nr_nodes_env1_old = len(grid_env.graph.all_nodes)
 
-        # create a new environment
-        new_boundary_coordinates = [(0.1, 0.1), (10.1, 0.1), (9.1, 5.1), (10.1, 10.1), (0.1, 10.1)]
-        new_list_of_holes = [[(3.1, 7.1), (5.1, 9.1), (4.6, 7.1), (5.1, 4.1), ], ]
-        new_environment = PolygonEnvironment()
-        # new_environment.store(new_boundary_coordinates, new_list_of_holes, validate=False)
-        new_environment.store(new_boundary_coordinates, new_list_of_holes, validate=True)
-        new_environment.prepare()
-        nr_nodes_env2 = len(new_environment.graph.all_nodes)
-        assert nr_nodes_env2 == 3, 'environment should have 3 extremities!'
+        poly_env = PolygonEnvironment()
+        poly_env.store(*POLY_ENV_PARAMS, validate=True)
+        poly_env.prepare()
+        nr_nodes_env2 = len(poly_env.graph.all_nodes)
 
-        nr_nodes_env1_new = len(environment.graph.all_nodes)
+        assert nr_nodes_env2 == 4, 'grid_env should have 4 extremities!'
+
+        nr_nodes_env1_new = len(grid_env.graph.all_nodes)
         assert nr_nodes_env1_new == nr_nodes_env1_old, \
-            'node amount of an environment should not change by creating another environment!'
-        assert environment.graph is not new_environment.graph, \
+            'node amount of an grid_env should not change by creating another grid_env!'
+        assert grid_env.graph is not poly_env.graph, \
             'different environments share the same graph object'
-        assert environment.graph.all_nodes is not new_environment.graph.all_nodes, \
+        assert grid_env.graph.all_nodes is not poly_env.graph.all_nodes, \
             'different environments share the same set of nodes'
+
+
+        try_test_cases(poly_env, TEST_DATA_POLY_ENV)
 
         # TODO test: When the paths should be blocked, use a single polygon with multiple identical
         #   vertices instead (also allowed).