+++ CHANGELOG.rst
@@ -7,7 +7,10 @@ Changelog
 
 * supporting only python 3.6+
 * fix #10: Memory leak in DirectedHeuristicGraph
+* fix BUG where extremities would be deleted from the visibility graph
 
+
+todo copy deep
 TODO sphinx documentation
 TODO api doc